Updates for ease of newcomers
Putting this here as a start to drafting up possible changes to this to make it easier for newcomers to jump in with helping with the site. This is to stay a draft until such a time we're satisfied with it. Plus this opened a issue we'll need to look at org wide for other operating systems. Toolbox is linux only as far as I can tell and I don't see any reasoning to change off that recommendation for linux, HOWEVER we will need alternatives that make it easier to follow our workflows and generate custom images imo.