Update index.md
Starting pr to update this info and link it here, this'll probably be a start in a move towards hopefully updating to a more advanced and proper site(especially getting around the warts of vuepress that we've noticed over time). But this info is necessary imo asap for package maintainers esp in a notice here depending on what we do going forward.